+++ test/unit/constraints/equal_to.jl
@@ -160,3 +160,17 @@ end
     @test CS.fix!(com, variables[2], 5; check_feasibility = false)
     @test !CS.is_constraint_violated(com, constraint, constraint.fct, constraint.set)
 end
+
+@testset "constraint without variables" begin
+    m = Model(optimizer_with_attributes(CS.Optimizer, "logging" => []))
+    @variable(m, -5 <= x[1:2] <= 5, Int)
+    @constraint(m, sum(0 .* x) == 10)
+    optimize!(m)
+    @test JuMP.termination_status(m) == MOI.INFEASIBLE
+
+    m = Model(optimizer_with_attributes(CS.Optimizer, "logging" => []))
+    @variable(m, -5 <= x[1:2] <= 5, Int)
+    @constraint(m, sum(0 .* x) == 0)
+    optimize!(m)
+    @test JuMP.termination_status(m) == MOI.OPTIMAL
+end